refactor: rename api endpoints
