+++ HubOneConfig.ts
@@ -1,7 +1,6 @@
 import type { Hub, Link, LinkGroup } from "@prisma/client";
 
 import type { FooterProps } from "./ui/components/Footer";
-import type { LinkGroupType } from "./ui/types/LinkGroupType";
 
 export interface HubOneConfigType {
   hub: Hub;
@@ -9,73 +8,73 @@ export interface HubOneConfigType {
   linkGroups: LinkGroup[];
 }
 
-const linkGroup0: LinkGroupType = {
-  title: "Link Group 1",
-  links: [
-    {
-      title: "1. Website",
-      description: "Description",
-      image: undefined,
-      link: "",
-      isInternal: true,
-    },
-    {
-      title: "2. Website",
-      description: "Description",
-      image: undefined,
-      link: "",
-      isInternal: true,
-    },
-    {
-      title: "3. Website",
-      description: "Description",
-      image: undefined,
-      link: "",
-      isInternal: true,
-    },
-  ],
-};
+// const linkGroup0: LinkGroupType = {
+//   title: "Link Group 1",
+//   links: [
+//     {
+//       title: "1. Website",
+//       description: "Description",
+//       image: undefined,
+//       link: "",
+//       isInternal: true,
+//     },
+//     {
+//       title: "2. Website",
+//       description: "Description",
+//       image: undefined,
+//       link: "",
+//       isInternal: true,
+//     },
+//     {
+//       title: "3. Website",
+//       description: "Description",
+//       image: undefined,
+//       link: "",
+//       isInternal: true,
+//     },
+//   ],
+// };
 
-const linkGroup1: LinkGroupType = {
-  title: "Link Group 2",
-  links: [
-    {
-      title: "1. Website",
-      description: "Description",
-      image: undefined,
-      link: "",
-      isInternal: true,
-    },
-    {
-      title: "2. Website",
-      description: "Description",
-      image: undefined,
-      link: "",
-      isInternal: true,
-    },
-    {
-      title: "3. Website",
-      description: "Description",
-      image: undefined,
-      link: "",
-      isInternal: true,
-    },
-    {
-      title: "4. Website",
-      description: "Description",
-      image: undefined,
-      link: "",
-      isInternal: false,
-    },
-    {
-      title: "5. Website",
-      description: "Description",
-      image: undefined,
-      link: "",
-      isInternal: false,
-    },
-  ],
-};
+// const linkGroup1: LinkGroupType = {
+//   title: "Link Group 2",
+//   links: [
+//     {
+//       title: "1. Website",
+//       description: "Description",
+//       image: undefined,
+//       link: "",
+//       isInternal: true,
+//     },
+//     {
+//       title: "2. Website",
+//       description: "Description",
+//       image: undefined,
+//       link: "",
+//       isInternal: true,
+//     },
+//     {
+//       title: "3. Website",
+//       description: "Description",
+//       image: undefined,
+//       link: "",
+//       isInternal: true,
+//     },
+//     {
+//       title: "4. Website",
+//       description: "Description",
+//       image: undefined,
+//       link: "",
+//       isInternal: false,
+//     },
+//     {
+//       title: "5. Website",
+//       description: "Description",
+//       image: undefined,
+//       link: "",
+//       isInternal: false,
+//     },
+//   ],
+// };
 
 // export const hubOneConfig: HubOneConfigType = {
 //   companyLogo: "/logo/hubone_logo.svg",
