+++ lib/requests/footerLink/getFooterLinks.ts
