+++ lib/requests/hub/getHubs.ts
@@ -3,7 +3,7 @@ import type { Hub } from "@prisma/client";
 import axios from "../../axios";
 
 export default async () => {
-  const { data } = await axios.get<Hub[]>("hub", {
+  const { data } = await axios.get<Hub[]>("hubs", {
     headers: {
       Accept: "application/json",
     },
