+++ lib/requests/linkGroup/getLinkGroups.ts
