+++ pages/api/v1/hubs/[id].ts
